+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## Sep 27, 2022 Update
+
+### Changes
+
+- Changed $userState.otherUser to $partnerToken, now only user being updated/read will be changed
+- Changed db schema to follow UserToken type (User.id => User.sub, first => given_name, last => family_name)
+
 ## Sep 23, 2022 Update
 
 ### Added